recommender/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.1+cpu
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
httpx==0.18.2